Create case information for offenders from England or Wales
Additional information is required for these offender's, this
includes whether the case is NPS or CRC, the offender's tier
and the team that the offender belongs to.

In this commit we are able to save case information for offenders
from England or Wales. Note: Offenders from England includes foreign
nationals also.

Show appropriate error messages when form not completed correctly